build: rename noreturn into __noreturn
Prior this change was noreturn used as a qualifier for functions
that never returns. This change replaces it with __noreturn.

The issue with label noreturn is that it is a reserved word used
for attribute qualifiers. Using __noreturn does not conflict with
this reserved word and easy porting of framework in over
environments such as the OP-TEE core firmware.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>

# Conflicts:
#	framework/include/internal/fwk_thread.h