prototyping SCMI server in optee
This change introduces OP-TEE as a product for SCP-firmware
generation. When built as such, SCP-firmware generates a
library which OP-TEE embeds in secure core as a SCMI server
stack.

This change introduces several flavors for the OP-TEE product,
depending on the target board.

Hack done in fwk_assert.c due to a build warning found when building
from OP-TEE source tree.

Signed-off-by: Vincent Guittot <vincent.guitto@linaro.org>
Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>