[OP-TEE hack] framework: Assign a thread context for message processing
When built for OP-TEE Core, SCP-firmware modified to assign a OP-TEE
thread context for an incoming message and ensure the message processing
is tied to that thread. This change original implementation in that
there can be several thread context whereas there was a only a single
possible processing context prior this change.

Signed-off-by: Etienne Carriere <etienne.carriere@linaro.org>
Signed-off-by: Vincent Guittot <vincent.guittot@linaro.org>